refactored player pages to use advquery method
