+++ routes/api.js
@@ -1,8 +1,7 @@
 var express = require('express');
 var api = express.Router();
 var constants = require('../constants');
-var db = require('../db');
-var queries = require('../queries');
+var advQuery = require('../advquery');
 api.get('/items', function(req, res) {
     res.json(constants.items[req.query.name]);
 });
@@ -11,26 +10,39 @@ api.get('/abilities', function(req, res) {
 });
 api.get('/matches', function(req, res, next) {
     var draw = Number(req.query.draw);
+    var select = req.query.select || {};
+    //api limits the fields that will be returned per match
+    //we add the player data if a player selection field present?
+    //todo right now this just always includes all the players
+    var project = {
+        start_time: 1,
+        match_id: 1,
+        cluster: 1,
+        game_mode: 1,
+        duration: 1,
+        parse_status: 1,
+        players: 1
+    };
     var start = Number(req.query.start);
     var limit = Number(req.query.length);
     //if limit is 0 or too big, reset it
-    //limit the number of results when using api to prevent abuse
+    //api limits the number of results it will return
+    //todo can we aggregate on large limit, but only return small limit?
+    //current implementation limits before aggregation
     limit = (!limit || limit > 10) ? 1 : limit;
-    var select = req.query.select || {};
     //var sort = makeSort(req.query.order, req.query.columns);
+    //api enforces sort by match_id
     var sort = {
         "match_id": -1
     };
-    var project = req.query.project || {};
-    if (select["players.account_id"]) {
-        //convert the passed account id to number
-        select["players.account_id"] = Number(select["players.account_id"]);
-    }
     advQuery(select, {
         limit: limit,
         skip: start,
         sort: sort,
-        fields: project
+        fields: project,
+        advQuery: {
+            filter: {}
+        }
     }, function(err, result) {
         if (err) {
             return next(err);
@@ -57,55 +69,4 @@ function makeSort(order, columns) {
     return sort;
 }
 */
-function advQuery(select, options, cb) {
-    //currently api is using this
-    //matches page, want matches fitting criteria
-    //player matches page, want winrate, matches fitting criteria
-    //player trends page, want agregation on matches fitting criteria
-    //custom query wants some fields back, and some criteria, with aggregation on those fields
-    //client options should include:
-    //filter: specific player/specific hero id
-    //filter: specific player was also in the game (use players.account_id with $and, but which player gets returned by projection?)
-    //filter: specific hero was played by me, was on my team, was against me, was in the game
-    //filter: specific game modes
-    //filter: specific patches
-    //filter: specific regions
-    //filter: detect no stats recorded (algorithmically)
-    //filter: significant game modes only    
-    //client calls api, which processes a maximum number of matches (currently 10, parsed matches are really big and we dont want to spend massive bandwidth!)
-    //can we increase the limit depending on the options passed?  if a user requests just a field or two we can return more
-    //use advquery function as a wrapper around db.matches.find to do processing that mongo can't
-    //select, a mongodb search hash
-    //options, a mongodb/monk options hash
-    //CONSTRAINT: each match can only have a SINGLE player matching the condition in order to make winrate/aggregations meaningful!
-    //check select.keys to see if user requested special conditions
-    //check options.fields.keys to see if user requested special fields, aggregate the selected fields
-    //we need to pass aggregator specific fields since not all fields may exist (since we projected)
-    //we can do indexes on the parsed data to enable mongo lookup, or post-process it in js
-    //fields (projection), limit, skip, sort (but sorts are probably best done in js)
-    //if selecting by account_id or hero_id, we project only that user in players array
-    //if (select["players.account_id"] || select["players.hero_id"]){options.fields["players.$"] = 1;}
-    /*
-    Enhance the /api/matches endpoint:
-Restrict projection per match to a very small set (to prevent bandwidth abuse)
-Set a limit on how many matches we will aggregate (max result set size)
-Set a limit on results returned at once to prevent bandwidth abuse
-Do aggregation on the results (Since aggData is relatively small)?
-Or we could just return winrate, count by iterating through the result set
-
-The problem with using datatables to interface with the api is that the query string suddenly gets a lot harder to build. 
-Maybe we can add jQuery helpers to construct the query and use it for both matches tab (lists matches, reports winrate fitting the advanced query conditions) and trends tabs (aggregates data fitting the advanced query conditions)
-*/
-    db.matches.find(select, options, function(err, matches) {
-        if (err) {
-            return cb(err);
-        }
-        //filter and send through aggregator?
-        var results = {
-            aggData: null,
-            data: matches
-        };
-        cb(err, results);
-    });
-}
 module.exports = api;
