+++ routes/matches.js
@@ -41,7 +41,7 @@ matches.get('/:match_id/:info?', function(req, res, next) {
     var info = matchPages[req.params.info] ? req.params.info : "index";
     console.timeEnd("match page");
     if (req.query.json){
-        return res.json(match);
+        //return res.json(match);
     }
     res.render("match_" + info, {
         route: info,
