+++ views/navbar.jade
@@ -14,6 +14,8 @@
         li
           a(href='/status') Status
         li
+          a(href="/request") Request
+        li
           a(href='/about') FAQ 
         li
           a(href='/page/1') Blog
